+++ src/weather_companion/weather_station/__init__.py
@@ -1,3 +1,3 @@
+from .base import WeatherStation, WeatherStationError
 from .location import Location
-from .open_weather_map import OpenWeatherMap, OpenWeatherMapClient
-from .station import WeatherStation, WeatherStationError
+from .open_weather_map import OpenWeatherMapClient, OWMWeatherStation
